requirements.txt
Flask==3.1.2
Flask-MySQLdb==2.0.0
mysqlclient==2.2.7
scikit-learn==1.3.2
numpy==1.26.2
google-generativeai==0.3.1
python-dotenv==1.0.0